requirements.txt
python-telegram-bot[job-queue]==21.6
python-dotenv==1.0.1
pytz==2024.1
asyncpg==0.29.0
